Move `terminal_key_paths()` to 'atomic_reactor/util.py'
